Update Dockerfile to use .NET SDK and Runtime 8.0
The Dockerfile has been updated to use .NET SDK 8.0 instead of .NET SDK 6.0 for building the application. This change is reflected in the `FROM` directive at the beginning of the build stage.

Additionally, the base image for the runtime environment has been updated from .NET Runtime 6.0 to .NET Runtime 8.0. This change is reflected in the `FROM` directive at the beginning of the runtime stage.

These updates ensure compatibility with the latest features and improvements in .NET, and may provide performance improvements and new features available in the latest runtime version.